feat: finish data generator
