+++ seeders/data.json
@@ -43,12 +43,6 @@
       "field": "樂天桃園棒球場",
       "orientation": "一壘側",
       "level": "下層",
-      "name": "東下C"
-    },
-    {
-      "field": "樂天桃園棒球場",
-      "orientation": "一壘側",
-      "level": "下層",
       "name": "東下G"
     },
     {
