+++ src/models/seat.js
@@ -1,0 +1,10 @@
+const prisma = require('../config/prisma');
+
+class SeatModel {
+  constructor() {}
+  async _truncate() {
+    await prisma.seats.deleteMany({});
+  }
+}
+
+module.exports = SeatModel;
