+++ src/models/space/constants.js
@@ -1,0 +1,6 @@
+const spaceTypeMap = {
+  seat: 'seat',
+  pillar: 'pillar',
+};
+
+module.exports = { spaceTypeMap };
