+++ src/services/photo-service.js
@@ -1,7 +1,7 @@
 const R = require('ramda');
 const BaseService = require('./base');
 const PhotoModel = require('../models/photo');
-const SpaceModel = require('../models/space/space');
+const SpaceModel = require('../models/space/index');
 const PrivateError = require('../errors/error/private-error');
 const GeneralError = require('../errors/error/general-error');
 const postPhotoErrorMap = require('../errors/post-photo-error');
